[pulsetracker-0001] Move error class to utils folder
