+++ packages/plugin/src/utils/error.ts
@@ -1,0 +1,8 @@
+export class ResponseError extends Error {
+  response: Response;
+
+  constructor(message: string, res: Response) {
+    super(message);
+    this.response = res;
+  }
+}
